Refactor replica scaling tests to deployment scaling with improved maintainability
- Renamed replica_scaling -> deployment_scaling for better terminology
- Simplified from 7+ complex tests to 2 core maintainable scenarios:
  * Verify_Pod_Scaling_With_YuniKorn_Scheduler (2->3->2 scaling with queue verification)
  * Verify_Preemption_During_Deployment_Scaling_Operations (priority-based preemption)
- Improved API server performance: polling interval 100ms -> 1s
- Enhanced code quality:
  * Consistent *v1.Pod pointer usage
  * Readable variable declarations (gracePeriod := int64(5))
  * Fixed linter warnings and improved type handling
- Made tests more robust with flexible YuniKorn allocation count checks
- All tests pass successfully with improved performance and maintainability
